Core Properties
Added a property class which acts like a reference but when it's value is read a function is called and when its value it set another function is called.
Added some missing properties to the core variant types, and added a specialization of property for each that also makes all of its properties and methods available!